# web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "SourceApplyResult": {
        "properties": {
          "action": {
            "enum": [
              "added",
              "rotated",
              "removed",
              "disabled",
              "unchanged"
            ],
            "type": "string"
          },
          "applied": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "persisted": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SourceRosterEntry": {
        "properties": {
          "component": {
            "readOnly": true,
            "type": "string"
          },
          "config": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "enabled": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poll_seconds": {
            "type": "integer"
          },
          "source_mode": {
            "enum": [
              "export",
              "live"
            ],
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SourceRosterInput": {
        "properties": {
          "config": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "enabled": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poll_seconds": {
            "type": "integer"
          },
          "tenant": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "tenant"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque session (olvs_) or API (olvk_) token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST contract for the Olivares AI control plane. Authentication is an opaque bearer token (session olvs_… or API key olvk_…). Tenant is resolved from a bound token or the X-Olivares-Tenant header.",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "title": "Olivares AI control plane API",
    "version": "v1",
    "x-stability-policy": "https://olivares.ai/docs"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/console/sources": {
      "delete": {
        "description": "Removes one configured source connector.",
        "operationId": "deleteSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRosterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceApplyResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict / setup required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove a source connector",
        "tags": [
          "console"
        ],
        "x-required-permission": "system:admin",
        "x-stability": "stable"
      },
      "get": {
        "description": "Lists the configured source connectors with the live status of each.",
        "operationId": "listSources",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "sources": {
                      "items": {
                        "$ref": "#/components/schemas/SourceRosterEntry"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict / setup required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List configured source connectors and their status",
        "tags": [
          "console"
        ],
        "x-required-permission": "system:admin",
        "x-stability": "stable"
      },
      "put": {
        "description": "Creates or replaces one source connector and reconfigures it live.",
        "operationId": "putSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRosterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceApplyResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict / setup required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create or update a source connector",
        "tags": [
          "console"
        ],
        "x-required-permission": "system:admin",
        "x-stability": "stable"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "this engine",
      "url": "/"
    }
  ],
...
}